Revamp WebStlink callback machinery
Callback events are now guaranteed to execute in order - if an 'inspect'
event is triggered before a 'halted' event, all of the 'inspect'
callbacks will have a chance to run before any of the 'halted' callbacks
are run.
Callbacks can stop other callbacks from firing on the same event by
returning an explicit false (not falsy) value.
'halted' and 'resumed' callbacks now consistently receive the CPU status
as a callback parameter